feat(poller): implementeer periodieke scheduler voor dagelijkse herinneringsmail
Breidt de bestaande Poller-app uit met een tweede geplande taak voor
het versturen van dagelijkse herinneringen aan behandelaars van verlopen
contactverzoeken (Feature #413).
