+++ InterneTaakAfhandeling.Poller/Features/IPollerJob.cs
@@ -1,0 +1,6 @@
+namespace InterneTaakAfhandeling.Poller.Features;
+
+public interface IPollerJob
+{
+    Task ExecuteAsync(CancellationToken cancellationToken = default);
+}
