+++ InterneTaakAfhandeling.Poller/Features/NieuweInternetaakNotificatie/NieuweInternetaakNotificatiePollerJob.cs
@@ -1,0 +1,7 @@
+namespace InterneTaakAfhandeling.Poller.Features.NieuweInternetaakNotificatie;
+
+public sealed class NieuweInternetaakNotificatiePollerJob(INieuweInternetakenProcessor processor) : IPollerJob
+{
+    public Task ExecuteAsync(CancellationToken cancellationToken = default)
+        => processor.NotifyAboutNewInternetakenAsync();
+}
