+++ charts/internetaakafhandeling/templates/nieuwe-internetaak-notificatie-cronjob.yaml
@@ -1,23 +1,23 @@
 apiVersion: batch/v1
 kind: CronJob
 metadata:
-  name: {{ include "internetaakafhandeling.poller.name" . }}
+  name: {{ include "internetaakafhandeling.nieuweInternetaakNotificatie.name" . }}
   labels:
-    {{- include "internetaakafhandeling.poller.labels" . | nindent 4 }}
+    {{- include "internetaakafhandeling.nieuweInternetaakNotificatie.labels" . | nindent 4 }}
 spec:
-  schedule: "{{ .Values.poller.schedule }}"
+  schedule: "{{ .Values.nieuweInternetaakNotificatie.schedule }}"
   jobTemplate:
     spec:
       template:
         metadata:
           labels:
-            {{- include "internetaakafhandeling.poller.selectorLabels" . | nindent 12 }}
+            {{- include "internetaakafhandeling.nieuweInternetaakNotificatie.selectorLabels" . | nindent 12 }}
         spec:
           containers:
-            - name: internetaakafhandeling-poller
-              image: "{{ .Values.poller.image.repository }}:{{ .Values.poller.image.tag }}"
-              imagePullPolicy: {{ .Values.poller.image.pullPolicy }}
-              {{- with .Values.poller.resources }}
+            - name: nieuwe-internetaak-notificatie
+              image: "{{ .Values.nieuweInternetaakNotificatie.image.repository }}:{{ .Values.nieuweInternetaakNotificatie.image.tag }}"
+              imagePullPolicy: {{ .Values.nieuweInternetaakNotificatie.image.pullPolicy }}
+              {{- with .Values.nieuweInternetaakNotificatie.resources }}
               resources:
                 {{- toYaml . | nindent 16 }}
               {{- end }}
